FEP-ADK-MCP/requirements.txt
google-adk
litellm
matplotlib
pandas
langchain
langchain_experimental
pyodbc
sentence-transformers
